move logic to agent
